+++ week4-homework/week4-homework/package.json
@@ -12,7 +12,8 @@
   "dependencies": {
     "@tailwindcss/vite": "^4.1.6",
     "react": "^19.1.0",
-    "react-dom": "^19.1.0"
+    "react-dom": "^19.1.0",
+    "react-router-dom": "^7.6.0"
   },
   "devDependencies": {
     "@eslint/js": "^9.25.0",
